Added sqrt

+++ src/test/java/info/dmerej/CalculatorTest.java
@@ -19,7 +19,7 @@ public class CalculatorTest {
 
     @ParameterizedTest
     @CsvFileSource(resources = "/test_values.csv", numLinesToSkip = 1)
-    void given_source_value_when_RPNEvaluation_return_good_result(String input, int expected) {
+    void given_source_value_when_RPNEvaluation_return_good_result(String input, double expected) {
         assertThat(
                 calculator.RPNEvaluation(input)
         ).isEqualTo(expected);
@@ -33,7 +33,7 @@ public class CalculatorTest {
         final String value = "1.3 5.5 +";
 
         //When
-        final Integer result = calculator.RPNEvaluation(value);
+        final Double result = calculator.RPNEvaluation(value);
 
         //then
         assertThat(result).isNull();
@@ -46,7 +46,7 @@ public class CalculatorTest {
         final String value = "5 +";
 
         //When
-        final Integer result = calculator.RPNEvaluation(value);
+        final Double result = calculator.RPNEvaluation(value);
 
         //then
         assertThat(result).isNull();
@@ -59,7 +59,7 @@ public class CalculatorTest {
         final String value = "1 + 3";
 
         //When
-        final Integer result = calculator.RPNEvaluation(value);
+        final Double result = calculator.RPNEvaluation(value);
 
         //then
         assertThat(result).isNull();
